Add post_body field to cfssl dict.
`post_body` overrides the `request` field, which is to say that if
`post_body` is present then `request` will be ignored, and if both are
present then `post_body` takes precedence.

Use two fields in this way because existing deployment logic specifies
`request`, and we don't want to break that.